CostFunc函数，mOnce * shape.k * FP16_SIZE可能发生整数溢出

Created-by: wmg1
Commit-by: wmg
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
改动原因：ascend-transformer-boost/src/kernels/kernels/matmul/tiling/pp_matmul_nz_tiling.cpp的函数GetTiling调用CostFunc内部时，计算表达式mOnce * shape.k * FP16_SIZE可能发生整数溢出。原因是未校验opShape.k的范围(最大可为UINT32_MAX)。
所采取的方法：对mOnce * shape.k * FP16_SIZE  升为uint64。
## checklist
<!--  [x] 表示选中 -->
- [ ] 是否通过本地IDE对代码进行静态检查.
- [ ] 是否通过本地IDE对代码进行格式化处理
- [x] 是否进行空指针校验
- [ ] 是否进行返回值校验
- [ ] 是否正确释放new/malloc申请的内存
- [ ] 是否充分考虑接口的异常场景
- [ ] 是否正确记录错误日志

See merge request: cann/ascend-transformer-boost!1657

+++ src/kernels/kernels/matmul/common/common_tiling.h
@@ -142,10 +142,10 @@ inline __attribute__((always_inline)) float CostFunc(const HardwareType &hwInfor
     uint32_t blockDim = std::min(coreNeed, hwInfor.coreNum);
     uint32_t mOnce = blockDim < nLoop ? shape.m0 : blockDim / nLoop * shape.m0;
     uint32_t nOnce = blockDim < nLoop ? hwInfor.coreNum * shape.n0 : shape.n;
-    if (mOnce * shape.k * FP16_SIZE > hwInfor.l2Size) {
+    if (static_cast<uint64_t>(mOnce) * shape.k * FP16_SIZE > static_cast<uint64_t>(hwInfor.l2Size)) {
         aCoef = bwCoef;
     }
-    if (nOnce * shape.k * FP16_SIZE > hwInfor.l2Size) {
+    if (static_cast<uint64_t>(nOnce) * shape.k * FP16_SIZE > static_cast<uint64_t>(hwInfor.l2Size)) {
         bCoef = bwCoef;
     }
     return 1 / (aCoef * static_cast<float>(shape.n0)) + 1 / (bCoef * static_cast<float>(shape.m0));
